Fixed the Prev/Next buttons by making the queue remove commands from its queue once it has done them. This also avoids the 100% cpu usage with it calling next constantly after the animation's finished.
Fixed the queue carrying on after you click restart by making
the shell pass these clicks to the panel, and adding a
stop method on the queue which interrupts the runner
thread. Went through the entire project removing most
instances of catch InterruptedException, because handling
this exception clears the Thread.interrupted() flag, and
prevents the queue from actually stopping. These are
passed up now, and the queue deals with them.

Ran ShellGraphAnimator through the code formatter.
Sorry Steven. The lack of whitespace was really
beginning to grate.

Might have inadvertently fixed fastForward() by making it
synchronised.

Fixed positioning of vectors when you roll the state back,
by copying the array of colsOccupied, rather than putting
a reference to it in the state object.